cleanup, little refactor and moved to async methods for writing to fs
